requirements.txt
Flask>=2.0
python-dotenv>=0.21
pytest>=7.0
